+++ src/propertyFields/collectionData/collectionNumberField/ICollectionNumberFieldProps.ts
@@ -3,6 +3,7 @@ import { ICustomCollectionField } from "..";
 export interface ICollectionNumberFieldProps {
   field: ICustomCollectionField;
   item: any;
+  disableEdit: boolean;
 
   fOnValueChange: (fieldId: string, value: any) => void;
   fValidation: (field: ICustomCollectionField, value: any) => Promise<string>;