+++ src/services/SPTermStorePickerService.ts
@@ -8,18 +8,14 @@
 import { SPHttpClient, SPHttpClientResponse, ISPHttpClientOptions } from '@microsoft/sp-http';
 import { Environment, EnvironmentType } from '@microsoft/sp-core-library';
 import { IWebPartContext } from '@microsoft/sp-webpart-base';
-import { IPropertyFieldTermPickerHostProps } from './../propertyFields/termPicker/IPropertyFieldTermPickerHost';
-import { ISPTermStores, ISPTermStore, ISPTermGroups, ISPTermGroup, IPickerTerms, IPickerTerm } from './../propertyFields/termPicker/IPropertyFieldTermPicker';
-import { ITermStore, ITerms, ITerm, IGroup, ITermSet, ISPTermStorePickerServiceProps } from './ISPTermStorePickerService';
+import { IPickerTerm } from './../propertyFields/termPicker/IPropertyFieldTermPicker';
+import { ITermStore, ITerms, ITerm, IGroup, ITermSet, ISPTermStorePickerServiceProps, ISPTermStorePickerService, TermStorePickerServiceHelper, ITermSets } from './ISPTermStorePickerService';
 import SPTermStoreMockHttpClient from './SPTermStorePickerMockService';
-import TermSet from './../propertyFields/termPicker/TermSet';
 
 /**
  * Service implementation to manage term stores in SharePoint
  */
-export default class SPTermStorePickerService {
-  private taxonomySession: string;
-  private formDigest: string;
+export default class SPTermStorePickerService implements ISPTermStorePickerService {
   private clientServiceUrl: string;
 
   /**
@@ -27,18 +23,6 @@ export default class SPTermStorePickerService {
    */
   constructor(private props: ISPTermStorePickerServiceProps, private context: IWebPartContext) {
     this.clientServiceUrl = this.context.pageContext.web.absoluteUrl + '/_vti_bin/client.svc/ProcessQuery';
-  }
-
-  /**
-   * Cleans the Guid from the Web Service response
-   * @param guid
-   */
-  public static cleanGuid(guid: string): string {
-    if (guid !== undefined) {
-      return guid.replace('/Guid(', '').replace('/', '').replace(')', '');
-    } else {
-      return '';
-    }
   }
 
   /**
@@ -81,7 +65,7 @@ export default class SPTermStorePickerService {
                 if (this.props.limitByTermsetNameOrID) {
                   const termsetNameOrId = this.props.limitByTermsetNameOrID;
                   termGroups = termGroups.map((group: IGroup) => {
-                    group.TermSets._Child_Items_ = group.TermSets._Child_Items_.filter((termSet: ITermSet) => termSet.Name === termsetNameOrId || SPTermStorePickerService.cleanGuid(termSet.Id).toLowerCase() === SPTermStorePickerService.cleanGuid(termsetNameOrId).toLowerCase());
+                    group.TermSets._Child_Items_ = group.TermSets._Child_Items_.filter((termSet: ITermSet) => termSet.Name === termsetNameOrId || TermStorePickerServiceHelper.cleanGuid(termSet.Id).toLowerCase() === TermStorePickerServiceHelper.cleanGuid(termsetNameOrId).toLowerCase());
                     return group;
                   });
                 }
@@ -155,13 +139,13 @@ export default class SPTermStorePickerService {
    * Retrieve all terms for the given term set
    * @param termsetId
    */
-  public async getAllTerms(termsetId: string): Promise<ITerm[]> {
+  public async getAllTerms(termSet: ITermSet): Promise<ITerm[]> {
     if (Environment.type === EnvironmentType.Local) {
       // If the running environment is local, load the data from the mock
       return this.getAllMockTerms();
     } else {
       // Request body to retrieve all terms for the given term set
-      const data = `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName=".NET Library" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><ObjectPath Id="30" ObjectPathId="29" /><Query Id="31" ObjectPathId="29"><Query SelectAllProperties="false"><Properties /></Query><ChildItemQuery SelectAllProperties="false"><Properties><Property Name="Name" ScalarProperty="true" /><Property Name="Id" ScalarProperty="true" /><Property Name="Description" ScalarProperty="true" /><Property Name="IsDeprecated" ScalarProperty="true" /><Property Name="IsAvailableForTagging" ScalarProperty="true" /><Property Name="IsRoot" ScalarProperty="true" /><Property Name="PathOfTerm" ScalarProperty="true" /><Property Name="TermSet" SelectAll="true" /><Property Name="Parent" SelectAll="true"><Query SelectAllProperties="false"><Properties /></Query></Property><Property Name="TermSet"><Query SelectAllProperties="false"><Properties><Property Name="Id" ScalarProperty="true" /></Properties></Query></Property></Properties></ChildItemQuery></Query></Actions><ObjectPaths><Method Id="29" ParentId="18" Name="GetAllTerms" /><Identity Id="18" Name="${termsetId}"
+      const data = `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName=".NET Library" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><ObjectPath Id="30" ObjectPathId="29" /><Query Id="31" ObjectPathId="29"><Query SelectAllProperties="false"><Properties /></Query><ChildItemQuery SelectAllProperties="false"><Properties><Property Name="Name" ScalarProperty="true" /><Property Name="Id" ScalarProperty="true" /><Property Name="Description" ScalarProperty="true" /><Property Name="IsDeprecated" ScalarProperty="true" /><Property Name="IsAvailableForTagging" ScalarProperty="true" /><Property Name="IsRoot" ScalarProperty="true" /><Property Name="PathOfTerm" ScalarProperty="true" /><Property Name="TermSet" SelectAll="true" /><Property Name="Parent" SelectAll="true"><Query SelectAllProperties="false"><Properties /></Query></Property><Property Name="TermSet"><Query SelectAllProperties="false"><Properties><Property Name="Id" ScalarProperty="true" /></Properties></Query></Property></Properties></ChildItemQuery></Query></Actions><ObjectPaths><Method Id="29" ParentId="18" Name="GetAllTerms" /><Identity Id="18" Name="${termSet._ObjectIdentity_}"
       /></ObjectPaths></Request>`;
 
       const reqHeaders = new Headers();
@@ -182,14 +166,14 @@ export default class SPTermStorePickerService {
             let terms = termStoreResult[0]._Child_Items_;
             // Clean the term ID and specify the path depth
             terms = terms.map(term => {
-              term.Id = SPTermStorePickerService.cleanGuid(term.Id);
+              term.Id = TermStorePickerServiceHelper.cleanGuid(term.Id);
               term['PathDepth'] = term.PathOfTerm.split(';').length;
               return term;
             });
             // Check if the term set was not empty
             if (terms.length > 0) {
               // Sort the terms by PathOfTerm
-              return terms.sort(this._sortTerms);
+              return terms.sort(TermStorePickerServiceHelper.sortTerms);
             }
           }
           return null;
@@ -217,6 +201,10 @@ export default class SPTermStorePickerService {
     }
   }
 
+  public async getGroupTermSets(group: IGroup): Promise<ITermSets> {
+    return group.TermSets;
+  }
+
   /**
      * Searches terms for the given term set
      * @param searchText
@@ -230,9 +218,12 @@ export default class SPTermStorePickerService {
       return new Promise<IPickerTerm[]>(resolve => {
         this.getTermStores().then(termStore => {
           let TermSetId = termSet;
-          if (!this.isGuid(termSet)) {
-            TermSetId = SPTermStorePickerService.cleanGuid(termStore[0].Groups._Child_Items_[0].TermSets._Child_Items_[0].Id);
+          if (!TermStorePickerServiceHelper.isGuid(termSet)) {
+            TermSetId = TermStorePickerServiceHelper.cleanGuid(termStore[0].Groups._Child_Items_[0].TermSets._Child_Items_[0].Id);
           }
+
+          const group: IGroup = this.getTermGroupByTermSetId(TermSetId, termStore);
+
           let data = `<Request xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="Javascript Library"><Actions><ObjectPath Id="456" ObjectPathId="455" /><ObjectIdentityQuery Id="457" ObjectPathId="455" /><ObjectPath Id="459" ObjectPathId="458" /><ObjectIdentityQuery Id="460" ObjectPathId="458" /><ObjectPath Id="462" ObjectPathId="461" /><ObjectIdentityQuery Id="463" ObjectPathId="461" /><ObjectPath Id="465" ObjectPathId="464" /><SetProperty Id="466" ObjectPathId="464" Name="TermLabel"><Parameter Type="String">${searchText}</Parameter></SetProperty><SetProperty Id="467" ObjectPathId="464" Name="DefaultLabelOnly"><Parameter Type="Boolean">true</Parameter></SetProperty><SetProperty Id="468" ObjectPathId="464" Name="StringMatchOption"><Parameter Type="Number">0</Parameter></SetProperty><SetProperty Id="469" ObjectPathId="464" Name="ResultCollectionSize"><Parameter Type="Number">10</Parameter></SetProperty><SetProperty Id="470" ObjectPathId="464" Name="TrimUnavailable"><Parameter Type="Boolean">true</Parameter></SetProperty><ObjectPath Id="472" ObjectPathId="471" /><Query Id="473" ObjectPathId="471"><Query SelectAllProperties="false"><Properties /></Query><ChildItemQuery SelectAllProperties="false"><Properties><Property Name="IsRoot" SelectAll="true" /><Property Name="Id" SelectAll="true" /><Property Name="Name" SelectAll="true" /><Property Name="PathOfTerm" SelectAll="true" /><Property Name="TermSet" SelectAll="true" /></Properties></ChildItemQuery></Query></Actions><ObjectPaths><StaticMethod Id="455" Name="GetTaxonomySession" TypeId="{981cbc68-9edc-4f8d-872f-71146fcbb84f}" /><Method Id="458" ParentId="455" Name="GetDefaultKeywordsTermStore" /><Method Id="461" ParentId="458" Name="GetTermSet"><Parameters><Parameter Type="Guid">${TermSetId}</Parameter></Parameters></Method><Constructor Id="464" TypeId="{61a1d689-2744-4ea3-a88b-c95bee9803aa}" /><Method Id="471" ParentId="461" Name="GetTerms"><Parameters><Parameter ObjectPathId="464" /></Parameters></Method></ObjectPaths></Request>`;
 
           const reqHeaders = new Headers();
@@ -257,12 +248,12 @@ export default class SPTermStorePickerService {
                 terms.forEach(term => {
                   if (term.Name.toLowerCase().indexOf(searchText.toLowerCase()) !== -1) {
                     returnTerms.push({
-                      key: SPTermStorePickerService.cleanGuid(term.Id),
+                      key: TermStorePickerServiceHelper.cleanGuid(term.Id),
                       name: term.Name,
                       path: term.PathOfTerm,
                       termSet: term.TermSet.Id,
-                      termSetName: term.TermSet.Name
-
+                      termSetName: term.TermSet.Name,
+                      termGroup: group ? TermStorePickerServiceHelper.cleanGuid(group.Id) : ''
                     });
                   }
                 });
@@ -289,7 +280,7 @@ export default class SPTermStorePickerService {
         this.getTermStores().then(termStore => {
           let termSetTerms: Array<Promise<ITerm[]>> = [];
           termStore[0].Groups._Child_Items_[0].TermSets._Child_Items_.forEach(ts => {
-            termSetTerms.push(this.getAllTerms(ts._ObjectIdentity_));
+            termSetTerms.push(this.getAllTerms(ts));
           });
 
           Promise.all(termSetTerms).then(results => {
@@ -299,11 +290,12 @@ export default class SPTermStorePickerService {
                 terms.forEach(term => {
                   if (term.Name.toLowerCase().indexOf(searchText.toLowerCase()) !== -1) {
                     returnTerms.push({
-                      key: SPTermStorePickerService.cleanGuid(term.Id),
+                      key: TermStorePickerServiceHelper.cleanGuid(term.Id),
                       name: term.Name,
                       path: term.PathOfTerm,
                       termSet: term.TermSet.Id,
-                      termSetName: term.TermSet.Name
+                      termSetName: term.TermSet.Name,
+                      termGroup: TermStorePickerServiceHelper.cleanGuid(termStore[0].Groups._Child_Items_[0].Id) // ??? I don't know why we always work with the 1st group
                     });
                   }
                 });
@@ -351,11 +343,12 @@ export default class SPTermStorePickerService {
               terms.forEach(term => {
                 if (term.Name.toLowerCase().indexOf(searchText.toLowerCase()) !== -1) {
                   returnTerms.push({
-                    key: SPTermStorePickerService.cleanGuid(term.Id),
+                    key: TermStorePickerServiceHelper.cleanGuid(term.Id),
                     name: term.Name,
                     path: term.PathOfTerm,
                     termSet: term.TermSet.Id,
-                    termSetName: term.TermSet.Name
+                    termSetName: term.TermSet.Name,
+                    termGroup: '' // didn't find a way to simply get Group in that situation. It won't affect functionality. Only switch between TermStorePicker and EnterpriseTermStorePicker
                   });
                 }
               });
@@ -366,26 +359,6 @@ export default class SPTermStorePickerService {
         });
       });
     }
-  }
-
-
-  private isGuid(strGuid: string): boolean {
-    return /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$/.test(strGuid);
-  }
-
-  /**
-   * Sort the terms by their path
-   * @param a term 2
-   * @param b term 2
-   */
-  private _sortTerms(a: ITerm, b: ITerm) {
-    if (a.PathOfTerm < b.PathOfTerm) {
-      return -1;
-    }
-    if (a.PathOfTerm > b.PathOfTerm) {
-      return 1;
-    }
-    return 0;
   }
 
   /**
@@ -404,5 +377,21 @@ export default class SPTermStorePickerService {
     return SPTermStoreMockHttpClient.getAllTerms().then((data) => {
       return data;
     }) as Promise<ITerm[]>;
+  }
+
+  private getTermGroupByTermSetId(termSetId: string, termStores: ITermStore[]): IGroup {
+    for (let i = 0, len = termStores.length; i < len; i++) {
+      const termStore = termStores[i];
+
+      for (let groupIdx = 0, groupsLen = termStore.Groups._Child_Items_.length; groupIdx < groupsLen; groupIdx++) {
+        const group = termStore.Groups._Child_Items_[groupIdx];
+
+        if (group.TermSets._Child_Items_.filter(ts => TermStorePickerServiceHelper.cleanGuid(ts.Id) === termSetId).length) {
+          return group;
+        }
+      }
+    }
+
+    return null;
   }
 }